Added much more robust and extensable menuing system and changed aproprite Start()'s into Awake()'s

+++ Platform Fighter/Assets/_SCRIPTS/NETWORKING/Lobby.cs
@@ -20,7 +20,7 @@ namespace NETWORKING
                 };
         }
 
-        public void JoinLobby(ulong id)
+        public void JoinLobby(int id)
         {
             
             Client.Instance.Lobby.Join(Client.Instance.Lobby.CurrentLobby);
@@ -44,7 +44,6 @@ namespace NETWORKING
                     Debug.Log($"Found Lobby: {lobby.Name}");
                 }
             };
-
         }
     }
 }
